Reorganize meta.ts
